requirements.txt
google-adk
google-cloud-storage
fastapi
uvicorn
python-dotenv